dev-requirements.txt
pytest
pytest-dotenv

# Bleeding edge
# git+https://github.com/dbt-labs/dbt-core.git@main#egg=dbt-tests-adapter&subdirectory=tests/adapter
# git+https://github.com/dbt-labs/dbt-core.git@main#egg=dbt-core&subdirectory=core
# git+https://github.com/dbt-labs/dbt-core.git@main#egg=dbt-postgres&subdirectory=plugins/postgres
# git+https://github.com/dbt-labs/dbt-redshift.git
# git+https://github.com/dbt-labs/dbt-snowflake.git
# git+https://github.com/dbt-labs/dbt-bigquery.git

# Most recent release candidates
dbt-tests-adapter==1.2.0
dbt-core==1.2.0
dbt-redshift==1.2.0
dbt-snowflake==1.2.0
dbt-bigquery==1.2.0